Feat/article bugfix

+++ news/views.py
@@ -5,14 +5,19 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 from django.shortcuts import get_object_or_404
-
+from django.db.models import Q 
 
 # 게시판 목록 기능 (누구나 이용 가능)
 class NewsListCreateView(APIView):
     permission_classes = [IsAuthenticatedOrReadOnly]
-
+    # 검색 목록 조회
     def get(self,request):
-        articles=Article.objects.all()
+        query = request.GET.get("search") if request.GET.get("search") else ""
+        articles=Article.objects.filter(
+                Q(title__icontains=query) |  # 제목에 검색어가 포함된 경우
+                Q(content__icontains=query) |  # 내용에 검색어가 포함된 경우
+                Q(author__username__icontains=query)  # 작성자 이름에 검색어가 포함된 경우
+            )
         serializer= ArticleSerializer(articles, many=True)
         return Response(serializer.data)
     
@@ -42,7 +47,6 @@ class NewsVote(APIView):
 class ArticleDetailView(APIView):
     permission_classes = [IsAuthenticatedOrReadOnly]
 
-
     # 게시글 상세 조회
     def get(self, request, news_id):
         article = get_object_or_404(Article, pk=news_id)
@@ -60,12 +64,18 @@ class ArticleDetailView(APIView):
             return Response(serializer.data)
 
 
-    # 게시글 삭제
+# 게시글 삭제
     def delete(self, request, news_id):
         article = get_object_or_404(Article, pk=news_id)
+        # 현재 요청한 사용자가 게시글의 작성자인지 확인
+        if article.author != request.user:
+            return Response("권한 없음", status=403)  # 403 Forbidden 응답 반환
+        
+        # 작성자인 경우 게시글을 삭제
         article.delete()
-        return Response(status=204)
 
+        # 삭제 완료 후 204 No Content 응답을 반환
+        return Response(status=204)
 
 #즐겨찾기
 class NewsFavorite(APIView):
@@ -100,9 +110,11 @@ class CommentViewSet(APIView):
             serializer.save(article=article, user=request.user)
             return Response(serializer.data, status=201)
 
+
+# 댓글 삭제
 class CommentDeleteViewSet(APIView):
     permission_classes = [IsAuthenticated]
-    # 댓글 삭제
+
     def delete(self, request, news_id, comment_id):
         comment = get_object_or_404(Comment, pk=comment_id)
         if comment.user != request.user:
@@ -135,3 +147,16 @@ class CommentFavorite(APIView):
         else:
             comment.favorite.add(request.user)
             return Response("즐겨찾기 완료!")
+
+
+# 대댓글
+class CommentReplyAPIView(APIView):
+    permission_classes = [IsAuthenticatedOrReadOnly]
+
+    # 대댓글 생성
+    def post(self, request, comment_id):
+        comment = get_object_or_404(Article, pk=comment_id)
+        serializer = CommentSerializer(data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            serializer.save(article=comment.article, user=request.user, parent_comment=comment)
+            return Response(serializer.data, status=201)
